added error checking in Look_for variable collection
May or may not work, don't have my Linux box up and running atm to check.  My goal is to pull all the blank space out of the variable when its collected to help with the searching function and also the file creation.  Users are silly, and I don't think writing "Don't use a space" will always help.